refactor: split app module
